put CHA code into rewrite.rs (most code is shared except for one function)
